Move route detection behind a capabilities facade
Prepare the ground for adding additional capability checks.  The created
ManagementClusterCapabilities struct can be passed to controller
creators as well as the controllers themselves for checks during
reconciles.
